requirements.txt
PyWavelets
opencv-python
seaborn
matplotlib
numpy
pandas
scikit-learn
flask_cors
flask
cv2